+++ packages/webview_flutter/webview_flutter_web/example/integration_test/legacy/webview_flutter_test.dart
@@ -3,12 +3,12 @@
 // found in the LICENSE file.
 
 import 'dart:async';
-import 'dart:html' as html;
 
 import 'package:flutter/material.dart';
 import 'package:flutter/widgets.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
+import 'package:web/helpers.dart';
 import 'package:webview_flutter_web_example/legacy/web_view.dart';
 
 void main() {
@@ -39,8 +39,8 @@ void main() {
     await controllerCompleter.future;
 
     // Assert an iframe has been rendered to the DOM with the correct src attribute.
-    final html.IFrameElement? element =
-        html.document.querySelector('iframe') as html.IFrameElement?;
+    final HTMLIFrameElement? element =
+        document.querySelector('iframe') as HTMLIFrameElement?;
     expect(element, isNotNull);
     expect(element!.src, primaryUrl);
   });
@@ -64,8 +64,8 @@ void main() {
     await controller.loadUrl(secondaryUrl);
 
     // Assert an iframe has been rendered to the DOM with the correct src attribute.
-    final html.IFrameElement? element =
-        html.document.querySelector('iframe') as html.IFrameElement?;
+    final HTMLIFrameElement? element =
+        document.querySelector('iframe') as HTMLIFrameElement?;
     expect(element, isNotNull);
     expect(element!.src, secondaryUrl);
   });