+++ packages/webview_flutter/webview_flutter_web/CHANGELOG.md
@@ -1,6 +1,7 @@
-## NEXT
+## 0.2.3
 
-* Updates minimum supported SDK version to Flutter 3.10/Dart 3.0.
+* Migrates to `package:web`.
+* Updates minimum supported SDK version to Flutter 3.16.0/Dart 3.2.0
 
 ## 0.2.2+3